refactor: refactoring gatsby node
